float answers, stable

+++ app/utils/count/count.pyi
@@ -3,7 +3,7 @@ def generate_question() -> dict: ...
 def check_winner(
     correct_res: float, 
     u1_id: int, 
-    u1_ans: int, 
+    u1_ans: float, 
     u2_id: int, 
-    u2_ans: int
+    u2_ans: float
 ) -> int: ...
